Macro for generic db tests
